python/requirements.txt
boto3
f5-icontrol-rest
